+++ src/index.ts
@@ -87,22 +87,27 @@ export interface Equivalence<S, T extends HKT, A> {
     g: (a: A) => a is U
   ) => Prism<S, Compose<T, F>, U>
   guard<U extends A>(g: (a: A) => a is U): Prism<S, Compose<T, Choice<A, U>>, U>
-  index(
-    i: number
-  ): IfElse<
-    Eq<A, string>,
-    Prism<S, Compose<T, DisallowTypeChange<string>>, string>,
-    Prism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
-  >
-  head(): IfElse<
-    Eq<A, string>,
-    Prism<S, Compose<T, DisallowTypeChange<string>>, string>,
-    Prism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
-  >
   find(
     predicate: (item: ElemType<A>) => boolean
   ): Prism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
   when(predicate: (item: A) => boolean): Prism<S, Compose<T, Union<A>>, A>
+
+  // Equivalence · RemovablePrism => RemovablePrism
+  compose<T2 extends HKT, A2>(
+    optic: RemovablePrism<A, T2, A2>
+  ): RemovablePrism<S, Compose<T, T2>, A2>
+  index(
+    i: number
+  ): IfElse<
+    Eq<A, string>,
+    RemovablePrism<S, Compose<T, DisallowTypeChange<string>>, string>,
+    RemovablePrism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+  >
+  head(): IfElse<
+    Eq<A, string>,
+    RemovablePrism<S, Compose<T, DisallowTypeChange<string>>, string>,
+    RemovablePrism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+  >
 
   // Equivalence · Traversal => Traversal
   compose<T2 extends HKT, A2>(
@@ -196,22 +201,27 @@ export interface Iso<S, T extends HKT, A> {
     g: (a: A) => a is U
   ) => Prism<S, Compose<T, F>, U>
   guard<U extends A>(g: (a: A) => a is U): Prism<S, Compose<T, Choice<A, U>>, U>
-  index(
-    i: number
-  ): IfElse<
-    Eq<A, string>,
-    Prism<S, Compose<T, DisallowTypeChange<string>>, string>,
-    Prism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
-  >
-  head(): IfElse<
-    Eq<A, string>,
-    Prism<S, Compose<T, DisallowTypeChange<string>>, string>,
-    Prism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
-  >
   find(
     predicate: (item: ElemType<A>) => boolean
   ): Prism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
   when(predicate: (item: A) => boolean): Prism<S, Compose<T, Union<A>>, A>
+
+  // Iso · RemovablePrism => RemovablePrism
+  compose<T2 extends HKT, A2>(
+    optic: RemovablePrism<A, T2, A2>
+  ): RemovablePrism<S, Compose<T, T2>, A2>
+  index(
+    i: number
+  ): IfElse<
+    Eq<A, string>,
+    RemovablePrism<S, Compose<T, DisallowTypeChange<string>>, string>,
+    RemovablePrism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+  >
+  head(): IfElse<
+    Eq<A, string>,
+    RemovablePrism<S, Compose<T, DisallowTypeChange<string>>, string>,
+    RemovablePrism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+  >
 
   // Iso · Traversal => Traversal
   compose<T2 extends HKT, A2>(
@@ -307,22 +317,27 @@ export interface Lens<S, T extends HKT, A> {
     g: (a: A) => a is U
   ) => Prism<S, Compose<T, F>, U>
   guard<U extends A>(g: (a: A) => a is U): Prism<S, Compose<T, Choice<A, U>>, U>
-  index(
-    i: number
-  ): IfElse<
-    Eq<A, string>,
-    Prism<S, Compose<T, DisallowTypeChange<string>>, string>,
-    Prism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
-  >
-  head(): IfElse<
-    Eq<A, string>,
-    Prism<S, Compose<T, DisallowTypeChange<string>>, string>,
-    Prism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
-  >
   find(
     predicate: (item: ElemType<A>) => boolean
   ): Prism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
   when(predicate: (item: A) => boolean): Prism<S, Compose<T, Union<A>>, A>
+
+  // Lens · RemovablePrism => RemovablePrism
+  compose<T2 extends HKT, A2>(
+    optic: RemovablePrism<A, T2, A2>
+  ): RemovablePrism<S, Compose<T, T2>, A2>
+  index(
+    i: number
+  ): IfElse<
+    Eq<A, string>,
+    RemovablePrism<S, Compose<T, DisallowTypeChange<string>>, string>,
+    RemovablePrism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+  >
+  head(): IfElse<
+    Eq<A, string>,
+    RemovablePrism<S, Compose<T, DisallowTypeChange<string>>, string>,
+    RemovablePrism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+  >
 
   // Lens · Traversal => Traversal
   compose<T2 extends HKT, A2>(
@@ -418,22 +433,27 @@ export interface Prism<S, T extends HKT, A> {
     g: (a: A) => a is U
   ) => Prism<S, Compose<T, F>, U>
   guard<U extends A>(g: (a: A) => a is U): Prism<S, Compose<T, Choice<A, U>>, U>
-  index(
-    i: number
-  ): IfElse<
-    Eq<A, string>,
-    Prism<S, Compose<T, DisallowTypeChange<string>>, string>,
-    Prism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
-  >
-  head(): IfElse<
-    Eq<A, string>,
-    Prism<S, Compose<T, DisallowTypeChange<string>>, string>,
-    Prism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
-  >
   find(
     predicate: (item: ElemType<A>) => boolean
   ): Prism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
   when(predicate: (item: A) => boolean): Prism<S, Compose<T, Union<A>>, A>
+
+  // Prism · RemovablePrism => RemovablePrism
+  compose<T2 extends HKT, A2>(
+    optic: RemovablePrism<A, T2, A2>
+  ): RemovablePrism<S, Compose<T, T2>, A2>
+  index(
+    i: number
+  ): IfElse<
+    Eq<A, string>,
+    RemovablePrism<S, Compose<T, DisallowTypeChange<string>>, string>,
+    RemovablePrism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+  >
+  head(): IfElse<
+    Eq<A, string>,
+    RemovablePrism<S, Compose<T, DisallowTypeChange<string>>, string>,
+    RemovablePrism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+  >
 
   // Prism · Traversal => Traversal
   compose<T2 extends HKT, A2>(
@@ -460,6 +480,122 @@ export interface Prism<S, T extends HKT, A> {
   compose<A2>(optic: Fold<A, A2>): Fold<S, A2>
 
   // Prism · Setter => Setter
+  compose<T2 extends HKT, A2>(
+    optic: Setter<A, T2, A2>
+  ): Setter<S, Compose<T, T2>, A2>
+  prependTo(): Setter<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+  appendTo(): Setter<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+}
+
+export interface RemovablePrism<S, T extends HKT, A> {
+  _tag: 'RemovablePrism'
+
+  // RemovablePrism · Equivalence => Prism
+  compose<T2 extends HKT, A2>(
+    optic: Equivalence<A, T2, A2>
+  ): Prism<S, Compose<T, T2>, A2>
+
+  // RemovablePrism · Iso => Prism
+  compose<T2 extends HKT, A2>(
+    optic: Iso<A, T2, A2>
+  ): Prism<S, Compose<T, T2>, A2>
+  iso<U>(
+    there: (a: A) => U,
+    back: (u: U) => A
+  ): Prism<S, Compose<T, Adapt<A, U>>, U>
+
+  // RemovablePrism · Lens => Prism
+  compose<T2 extends HKT, A2>(
+    optic: Lens<A, T2, A2>
+  ): Prism<S, Compose<T, T2>, A2>
+  prop<K extends keyof A>(key: K): Prism<S, Compose<T, Prop<A, K>>, A[K]>
+  path<
+    K1 extends keyof A,
+    K2 extends keyof A[K1],
+    K3 extends keyof A[K1][K2],
+    K4 extends keyof A[K1][K2][K3],
+    K5 extends keyof A[K1][K2][K3][K4]
+  >(
+    path: [K1, K2, K3, K4, K5]
+  ): Prism<S, Compose<T, Path5<A, K1, K2, K3, K4, K5>>, A[K1][K2][K3][K4][K5]>
+  path<
+    K1 extends keyof A,
+    K2 extends keyof A[K1],
+    K3 extends keyof A[K1][K2],
+    K4 extends keyof A[K1][K2][K3]
+  >(
+    path: [K1, K2, K3, K4]
+  ): Prism<S, Compose<T, Path4<A, K1, K2, K3, K4>>, A[K1][K2][K3][K4]>
+  path<K1 extends keyof A, K2 extends keyof A[K1], K3 extends keyof A[K1][K2]>(
+    path: [K1, K2, K3]
+  ): Prism<S, Compose<T, Path3<A, K1, K2, K3>>, A[K1][K2][K3]>
+  path<K1 extends keyof A, K2 extends keyof A[K1]>(
+    path: [K1, K2]
+  ): Prism<S, Compose<T, Path2<A, K1, K2>>, A[K1][K2]>
+  path<K1 extends keyof A>(path: [K1]): Prism<S, Compose<T, Prop<A, K1>>, A[K1]>
+  pick<K extends keyof A>(
+    keys: K[]
+  ): Prism<S, Compose<T, Plant<A, K>>, Pick<A, K>>
+  filter(
+    predicate: (item: ElemType<A>) => boolean
+  ): Prism<S, Compose<T, Union<A>>, A>
+
+  // RemovablePrism · Prism => Prism
+  compose<T2 extends HKT, A2>(
+    optic: Prism<A, T2, A2>
+  ): Prism<S, Compose<T, T2>, A2>
+  optional(): Prism<S, Compose<T, Optional>, Exclude<A, undefined>>
+  guard_<F extends HKT>(): <U extends A>(
+    g: (a: A) => a is U
+  ) => Prism<S, Compose<T, F>, U>
+  guard<U extends A>(g: (a: A) => a is U): Prism<S, Compose<T, Choice<A, U>>, U>
+  find(
+    predicate: (item: ElemType<A>) => boolean
+  ): Prism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+  when(predicate: (item: A) => boolean): Prism<S, Compose<T, Union<A>>, A>
+
+  // RemovablePrism · RemovablePrism => RemovablePrism
+  compose<T2 extends HKT, A2>(
+    optic: RemovablePrism<A, T2, A2>
+  ): RemovablePrism<S, Compose<T, T2>, A2>
+  index(
+    i: number
+  ): IfElse<
+    Eq<A, string>,
+    RemovablePrism<S, Compose<T, DisallowTypeChange<string>>, string>,
+    RemovablePrism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+  >
+  head(): IfElse<
+    Eq<A, string>,
+    RemovablePrism<S, Compose<T, DisallowTypeChange<string>>, string>,
+    RemovablePrism<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+  >
+
+  // RemovablePrism · Traversal => Traversal
+  compose<T2 extends HKT, A2>(
+    optic: Traversal<A, T2, A2>
+  ): Traversal<S, Compose<T, T2>, A2>
+  elems(): Traversal<S, Compose<T, Elems>, ElemType<A>>
+  chars(): RequireString<
+    A,
+    Traversal<S, Compose<T, DisallowTypeChange<string>>, string>
+  >
+  words(): RequireString<
+    A,
+    Traversal<S, Compose<T, DisallowTypeChange<string>>, string>
+  >
+
+  // RemovablePrism · Getter => AffineFold
+  compose<A2>(optic: Getter<A, A2>): AffineFold<S, A2>
+  to<B>(f: (a: A) => B): AffineFold<S, B>
+
+  // RemovablePrism · AffineFold => AffineFold
+  compose<A2>(optic: AffineFold<A, A2>): AffineFold<S, A2>
+
+  // RemovablePrism · Fold => Fold
+  compose<A2>(optic: Fold<A, A2>): Fold<S, A2>
+
+  // RemovablePrism · Setter => Setter
   compose<T2 extends HKT, A2>(
     optic: Setter<A, T2, A2>
   ): Setter<S, Compose<T, T2>, A2>
@@ -537,6 +673,15 @@ export interface Traversal<S, T extends HKT, A> {
   guard<U extends A>(
     g: (a: A) => a is U
   ): Traversal<S, Compose<T, Choice<A, U>>, U>
+  find(
+    predicate: (item: ElemType<A>) => boolean
+  ): Traversal<S, Compose<T, ElemUnion<A>>, ElemType<A>>
+  when(predicate: (item: A) => boolean): Traversal<S, Compose<T, Union<A>>, A>
+
+  // Traversal · RemovablePrism => Traversal
+  compose<T2 extends HKT, A2>(
+    optic: RemovablePrism<A, T2, A2>
+  ): Traversal<S, Compose<T, T2>, A2>
   index(
     i: number
   ): IfElse<
@@ -549,10 +694,6 @@ export interface Traversal<S, T extends HKT, A> {
     Traversal<S, Compose<T, DisallowTypeChange<string>>, string>,
     Traversal<S, Compose<T, ElemUnion<A>>, ElemType<A>>
   >
-  find(
-    predicate: (item: ElemType<A>) => boolean
-  ): Traversal<S, Compose<T, ElemUnion<A>>, ElemType<A>>
-  when(predicate: (item: A) => boolean): Traversal<S, Compose<T, Union<A>>, A>
 
   // Traversal · Traversal => Traversal
   compose<T2 extends HKT, A2>(
@@ -633,6 +774,13 @@ export interface Getter<S, A> {
     g: (a: A) => a is U
   ) => AffineFold<S, U>
   guard<U extends A>(g: (a: A) => a is U): AffineFold<S, U>
+  find(predicate: (item: ElemType<A>) => boolean): AffineFold<S, ElemType<A>>
+  when(predicate: (item: A) => boolean): AffineFold<S, A>
+
+  // Getter · RemovablePrism => AffineFold
+  compose<T2 extends HKT, A2>(
+    optic: RemovablePrism<A, T2, A2>
+  ): AffineFold<S, A2>
   index(
     i: number
   ): IfElse<Eq<A, string>, AffineFold<S, string>, AffineFold<S, ElemType<A>>>
@@ -641,8 +789,6 @@ export interface Getter<S, A> {
     AffineFold<S, string>,
     AffineFold<S, ElemType<A>>
   >
-  find(predicate: (item: ElemType<A>) => boolean): AffineFold<S, ElemType<A>>
-  when(predicate: (item: A) => boolean): AffineFold<S, A>
 
   // Getter · Traversal => Fold
   compose<T2 extends HKT, A2>(optic: Traversal<A, T2, A2>): Fold<S, A2>
@@ -708,6 +854,13 @@ export interface AffineFold<S, A> {
     g: (a: A) => a is U
   ) => AffineFold<S, U>
   guard<U extends A>(g: (a: A) => a is U): AffineFold<S, U>
+  find(predicate: (item: ElemType<A>) => boolean): AffineFold<S, ElemType<A>>
+  when(predicate: (item: A) => boolean): AffineFold<S, A>
+
+  // AffineFold · RemovablePrism => AffineFold
+  compose<T2 extends HKT, A2>(
+    optic: RemovablePrism<A, T2, A2>
+  ): AffineFold<S, A2>
   index(
     i: number
   ): IfElse<Eq<A, string>, AffineFold<S, string>, AffineFold<S, ElemType<A>>>
@@ -716,8 +869,6 @@ export interface AffineFold<S, A> {
     AffineFold<S, string>,
     AffineFold<S, ElemType<A>>
   >
-  find(predicate: (item: ElemType<A>) => boolean): AffineFold<S, ElemType<A>>
-  when(predicate: (item: A) => boolean): AffineFold<S, A>
 
   // AffineFold · Traversal => Fold
   compose<T2 extends HKT, A2>(optic: Traversal<A, T2, A2>): Fold<S, A2>
@@ -781,10 +932,13 @@ export interface Fold<S, A> {
   optional(): Fold<S, Exclude<A, undefined>>
   guard_<F extends HKT>(): <U extends A>(g: (a: A) => a is U) => Fold<S, U>
   guard<U extends A>(g: (a: A) => a is U): Fold<S, U>
-  index(i: number): IfElse<Eq<A, string>, Fold<S, string>, Fold<S, ElemType<A>>>
-  head(): IfElse<Eq<A, string>, Fold<S, string>, Fold<S, ElemType<A>>>
   find(predicate: (item: ElemType<A>) => boolean): Fold<S, ElemType<A>>
   when(predicate: (item: A) => boolean): Fold<S, A>
+
+  // Fold · RemovablePrism => Fold
+  compose<T2 extends HKT, A2>(optic: RemovablePrism<A, T2, A2>): Fold<S, A2>
+  index(i: number): IfElse<Eq<A, string>, Fold<S, string>, Fold<S, ElemType<A>>>
+  head(): IfElse<Eq<A, string>, Fold<S, string>, Fold<S, ElemType<A>>>
 
   // Fold · Traversal => Fold
   compose<T2 extends HKT, A2>(optic: Traversal<A, T2, A2>): Fold<S, A2>
@@ -827,6 +981,11 @@ export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
   optic1: Equivalence<S, T, A>,
   optic2: Prism<A, T2, A2>
 ): Prism<S, Compose<T, T2>, A2>
+// Equivalence · RemovablePrism => RemovablePrism
+export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
+  optic1: Equivalence<S, T, A>,
+  optic2: RemovablePrism<A, T2, A2>
+): RemovablePrism<S, Compose<T, T2>, A2>
 // Equivalence · Traversal => Traversal
 export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
   optic1: Equivalence<S, T, A>,
@@ -872,6 +1031,11 @@ export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
   optic1: Iso<S, T, A>,
   optic2: Prism<A, T2, A2>
 ): Prism<S, Compose<T, T2>, A2>
+// Iso · RemovablePrism => RemovablePrism
+export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
+  optic1: Iso<S, T, A>,
+  optic2: RemovablePrism<A, T2, A2>
+): RemovablePrism<S, Compose<T, T2>, A2>
 // Iso · Traversal => Traversal
 export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
   optic1: Iso<S, T, A>,
@@ -917,6 +1081,11 @@ export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
   optic1: Lens<S, T, A>,
   optic2: Prism<A, T2, A2>
 ): Prism<S, Compose<T, T2>, A2>
+// Lens · RemovablePrism => RemovablePrism
+export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
+  optic1: Lens<S, T, A>,
+  optic2: RemovablePrism<A, T2, A2>
+): RemovablePrism<S, Compose<T, T2>, A2>
 // Lens · Traversal => Traversal
 export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
   optic1: Lens<S, T, A>,
@@ -962,6 +1131,11 @@ export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
   optic1: Prism<S, T, A>,
   optic2: Prism<A, T2, A2>
 ): Prism<S, Compose<T, T2>, A2>
+// Prism · RemovablePrism => RemovablePrism
+export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
+  optic1: Prism<S, T, A>,
+  optic2: RemovablePrism<A, T2, A2>
+): RemovablePrism<S, Compose<T, T2>, A2>
 // Prism · Traversal => Traversal
 export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
   optic1: Prism<S, T, A>,
@@ -987,6 +1161,56 @@ export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
   optic1: Prism<S, T, A>,
   optic2: Setter<A, T2, A2>
 ): Setter<S, Compose<T, T2>, A2>
+// RemovablePrism · Equivalence => Prism
+export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
+  optic1: RemovablePrism<S, T, A>,
+  optic2: Equivalence<A, T2, A2>
+): Prism<S, Compose<T, T2>, A2>
+// RemovablePrism · Iso => Prism
+export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
+  optic1: RemovablePrism<S, T, A>,
+  optic2: Iso<A, T2, A2>
+): Prism<S, Compose<T, T2>, A2>
+// RemovablePrism · Lens => Prism
+export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
+  optic1: RemovablePrism<S, T, A>,
+  optic2: Lens<A, T2, A2>
+): Prism<S, Compose<T, T2>, A2>
+// RemovablePrism · Prism => Prism
+export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
+  optic1: RemovablePrism<S, T, A>,
+  optic2: Prism<A, T2, A2>
+): Prism<S, Compose<T, T2>, A2>
+// RemovablePrism · RemovablePrism => RemovablePrism
+export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
+  optic1: RemovablePrism<S, T, A>,
+  optic2: RemovablePrism<A, T2, A2>
+): RemovablePrism<S, Compose<T, T2>, A2>
+// RemovablePrism · Traversal => Traversal
+export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
+  optic1: RemovablePrism<S, T, A>,
+  optic2: Traversal<A, T2, A2>
+): Traversal<S, Compose<T, T2>, A2>
+// RemovablePrism · Getter => AffineFold
+export function compose<S, T extends HKT, A, A2>(
+  optic1: RemovablePrism<S, T, A>,
+  optic2: Getter<A, A2>
+): AffineFold<S, A2>
+// RemovablePrism · AffineFold => AffineFold
+export function compose<S, T extends HKT, A, A2>(
+  optic1: RemovablePrism<S, T, A>,
+  optic2: AffineFold<A, A2>
+): AffineFold<S, A2>
+// RemovablePrism · Fold => Fold
+export function compose<S, T extends HKT, A, A2>(
+  optic1: RemovablePrism<S, T, A>,
+  optic2: Fold<A, A2>
+): Fold<S, A2>
+// RemovablePrism · Setter => Setter
+export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
+  optic1: RemovablePrism<S, T, A>,
+  optic2: Setter<A, T2, A2>
+): Setter<S, Compose<T, T2>, A2>
 // Traversal · Equivalence => Traversal
 export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
   optic1: Traversal<S, T, A>,
@@ -1006,6 +1230,11 @@ export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
 export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
   optic1: Traversal<S, T, A>,
   optic2: Prism<A, T2, A2>
+): Traversal<S, Compose<T, T2>, A2>
+// Traversal · RemovablePrism => Traversal
+export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
+  optic1: Traversal<S, T, A>,
+  optic2: RemovablePrism<A, T2, A2>
 ): Traversal<S, Compose<T, T2>, A2>
 // Traversal · Traversal => Traversal
 export function compose<S, T extends HKT, A, T2 extends HKT, A2>(
@@ -1052,6 +1281,11 @@ export function compose<S, A, T2 extends HKT, A2>(
   optic1: Getter<S, A>,
   optic2: Prism<A, T2, A2>
 ): AffineFold<S, A2>
+// Getter · RemovablePrism => AffineFold
+export function compose<S, A, T2 extends HKT, A2>(
+  optic1: Getter<S, A>,
+  optic2: RemovablePrism<A, T2, A2>
+): AffineFold<S, A2>
 // Getter · Traversal => Fold
 export function compose<S, A, T2 extends HKT, A2>(
   optic1: Getter<S, A>,
@@ -1092,6 +1326,11 @@ export function compose<S, A, T2 extends HKT, A2>(
   optic1: AffineFold<S, A>,
   optic2: Prism<A, T2, A2>
 ): AffineFold<S, A2>
+// AffineFold · RemovablePrism => AffineFold
+export function compose<S, A, T2 extends HKT, A2>(
+  optic1: AffineFold<S, A>,
+  optic2: RemovablePrism<A, T2, A2>
+): AffineFold<S, A2>
 // AffineFold · Traversal => Fold
 export function compose<S, A, T2 extends HKT, A2>(
   optic1: AffineFold<S, A>,
@@ -1131,6 +1370,11 @@ export function compose<S, A, T2 extends HKT, A2>(
 export function compose<S, A, T2 extends HKT, A2>(
   optic1: Fold<S, A>,
   optic2: Prism<A, T2, A2>
+): Fold<S, A2>
+// Fold · RemovablePrism => Fold
+export function compose<S, A, T2 extends HKT, A2>(
+  optic1: Fold<S, A>,
+  optic2: RemovablePrism<A, T2, A2>
 ): Fold<S, A2>
 // Fold · Traversal => Fold
 export function compose<S, A, T2 extends HKT, A2>(
@@ -1175,13 +1419,22 @@ export function get<S, A>(
 }
 
 export function preview<S, A>(
-  optic: Prism<S, any, A> | Traversal<S, any, A> | AffineFold<S, A> | Fold<S, A>
+  optic:
+    | Prism<S, any, A>
+    | RemovablePrism<S, any, A>
+    | Traversal<S, any, A>
+    | AffineFold<S, A>
+    | Fold<S, A>
 ): (source: S) => A | undefined {
   return source => I.preview((optic as any)._ref, source)
 }
 
 export function collect<S, A>(
-  optic: Prism<S, any, A> | Traversal<S, any, A> | Fold<S, A>
+  optic:
+    | Prism<S, any, A>
+    | RemovablePrism<S, any, A>
+    | Traversal<S, any, A>
+    | Fold<S, A>
 ): (source: S) => A[] {
   return source => I.collect((optic as any)._ref, source)
 }
@@ -1192,6 +1445,7 @@ export function modify<S, T extends HKT, A>(
     | Iso<S, T, A>
     | Lens<S, T, A>
     | Prism<S, T, A>
+    | RemovablePrism<S, T, A>
     | Traversal<S, T, A>
 ): <B>(f: (a: A) => B) => (source: S) => Simplify<S, Apply<T, B>> {
   return f => source => I.modify((optic as any)._ref, f, source)
@@ -1203,10 +1457,17 @@ export function set<S, T extends HKT, A>(
     | Iso<S, T, A>
     | Lens<S, T, A>
     | Prism<S, T, A>
+    | RemovablePrism<S, T, A>
     | Traversal<S, T, A>
     | Setter<S, T, A>
 ): <B>(value: B) => (source: S) => Simplify<S, Apply<T, B>> {
   return value => source => I.set((optic as any)._ref, value, source)
+}
+
+export function remove<S>(
+  optic: RemovablePrism<S, any, any>
+): (source: S) => S {
+  return source => I.remove((optic as any)._ref, source)
 }
 
 // Taken from fp-ts